Linstor: Fix migrate primary storage
Always use RAW as copy sourceFormat for Linstor devices,
as all images are stored as RAW on Linstor, but we have to
pretend to use qcow2 to be able to use qcow2 as snapshot format.
Also use .getPath() for resource names, as they are altered after
a primary storage migration.